requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
deta==0.9
ecdsa==0.17.0
fastapi==0.65.3
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
idna==2.10
iniconfig==1.1.1
install==1.3.4
packaging==21.0
passlib==1.7.4
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.4
python-decouple==3.4
python-jose==3.3.0
python-multipart==0.0.5
requests==2.25.1
rsa==4.7.2
six==1.16.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
